Starting Enters membership through TouchScreen use case
Minor modifications for which classes I plan to use. There is a new package ending in ".gui"
One class for implementing the logic of entering the membeship number through a touchscreen
Another one that will test it.
Some Changes to SelfCheckooutSoftware to attach the touchscreen observer to screen device.
Added a constructor to TouchScreenSoftware so that it can take a SelfCheckoutStation.