ReceiptPrint Edits
I added changes to attendant actions, station, and receipt print. Now the attendant can select fill ink and paper in actions for a given station. Testing runs, and works 6/6 for tests. The testing finds if the ink or paper will throw a warning if empty, and will also test if attendant actions work.