Print Member in Command Line Receipt Test
I left out the "End of Receipt" part on the actual receipt, as it was running out of room for that line (not sure how to add an extra line).
It's included on the command line receipt to differentiate from all other stuff that gets printed to the command line.